tweaks to the integration tests/test building
The time taken to run all of the integration tests made the process of
making changes to the library painful. Although there aren't a huge
number of integration tests, each test is run multiple times, with
different transports and forceTLS enabled/disabled.

In order to speed up development feedback loops a slimmer integration
test suite is included which only tests the `ws` protocol. The slimmer
suite can be run by setting the env var MINIMAL_INTEGRATION_TESTS

While implementing the slimmer test suite a few other peculiarities were
addressed with the tests/build process. Many peculariaties remain.

- standardise options object for transport specific integration tests
- replace weird node_modules module
- build test configs in index and inject into tests to allow for easier
transport filtering

- build tests in development mode[0]
[0]: the default build uses 'production' mode, which minifies source. This
leads to helpful error messages like: <error> at
http://localhost:9876/base/spec/javascripts/integration/index.web.js?d6cc0b44ae6eb996ce8067ca46f80172a4750ea8:7
using development mode gives the slightly more helpful: <error> at at
webpack:///./src/core/strategies/strategy_builder.ts?:15
