+++ spec/javascripts/helpers/pusher_integration.ts
@@ -10,5 +10,4 @@ class PusherIntegration extends Pusher {
     )}
 
 }
-
 module.exports = PusherIntegration;
